requirements.txt
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.6.1
openai-whisper==20231117
httpx==0.26.0